requirements.txt
blinker==1.9.0
click==8.2.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
pandas==2.3.1
parse==1.20.2
pillow==11.3.0
pycocotools==2.0.10
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==15.0.1
Werkzeug==3.1.3
